% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_outgroup_to_phylogeny.R
\name{add_outgroup_to_phylogeny}
\alias{add_outgroup_to_phylogeny}
\title{Adds an outgroup to a phylogeny}
\usage{
add_outgroup_to_phylogeny(phylogeny, stem_length, outgroup_name = "Outgroup")
}
\arguments{
\item{phylogeny}{a phylogeny}

\item{stem_length}{How long will the length of the outgroup be before the crown?}

\item{outgroup_name}{How is the outgroup named?}
}
\value{
A phylogeny, of type phylo
}
\description{
Adds an outgroup to a phylogeny
}
\author{
Richel Bilderbeek
}

