% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_identical_phylogenies.R
\name{are_identical_phylogenies}
\alias{are_identical_phylogenies}
\title{Determines if two phylogenies are equal
as parsed by olli's rBEAST package its function beast2out.read.trees}
\usage{
are_identical_phylogenies(p, q)
}
\arguments{
\item{p}{the first phylogeny}

\item{q}{the second phylogeny}
}
\value{
TRUE or FALSE
}
\description{
Determines if two phylogenies are equal
as parsed by olli's rBEAST package its function beast2out.read.trees
}
\examples{
  p <- ape::rcoal(10)
  q <- ape::rcoal(10)
  testit::assert(are_identical_phylogenies(p, p))
  testit::assert(!are_identical_phylogenies(p, q))
}
\author{
Richel Bilderbeek
}

