% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phylogeny.R
\name{check_phylogeny}
\alias{check_phylogeny}
\title{Checks if the input is a phylogeny}
\usage{
check_phylogeny(phylogeny)
}
\arguments{
\item{phylogeny}{object to be checked}
}
\description{
Will stop if not, will do nothing otherwise
}
\author{
Richèl J.C. Bilderbeek
}
