% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_rbind_dfs.R
\name{create_random_df}
\alias{create_random_df}
\title{Creates a data frame with three columns:
(1) filename, a string, (2) replicate, a factor,
(3) value, a double. The number of rows is between
one and ten}
\usage{
create_random_df()
}
\description{
Creates a data frame with three columns:
(1) filename, a string, (2) replicate, a factor,
(3) value, a double. The number of rows is between
one and ten
}

