% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_rbind_dfs.R
\name{create_random_dfs}
\alias{create_random_dfs}
\title{Creates 100 data frames with three columns:
(1) filename, a string, (2) replicate, a factor,
(3) value, a double. The number of rows for each
data frame is between
one and ten}
\usage{
create_random_dfs(n = 100)
}
\description{
Creates 100 data frames with three columns:
(1) filename, a string, (2) replicate, a factor,
(3) value, a double. The number of rows for each
data frame is between
one and ten
}

