% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_nltt.R
\name{get_average_nltt}
\alias{get_average_nltt}
\title{Get the average nLTT from a collection of phylogenies}
\usage{
get_average_nltt(phylogenies, dt = 0.001, plot_nltts = FALSE,
  xlab = "Normalized Time", ylab = "Normalized Lineages", replot = FALSE,
  ...)
}
\arguments{
\item{phylogenies}{the phylogenies, where the phylogenies are of type 'phylo'}

\item{dt}{The timestep resolution, where 1/dt is the number of points evaluated}
}
\value{
Nothing
}
\description{
Get the average nLTT from a collection of phylogenies
}
\examples{
  get_average_nllt(rcoal(10))
  get_average_nllt(rcoal(10), dt = 0.1)
}

