% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_nltt_matrix.R
\name{get_average_nltt_matrix_impl_3}
\alias{get_average_nltt_matrix_impl_3}
\title{New function that gives tidy data}
\usage{
get_average_nltt_matrix_impl_3(phylogenies, dt)
}
\arguments{
\item{phylogenies}{the phylogenies, supplied as either a list or a multiPhylo object, where the phylogenies are of type 'phylo'}

\item{dt}{The timestep resolution, where 1/dt is the number of points evaluated}
}
\value{
A matrix of timepoints with the average number of (normalized) lineages through (normalized) time
}
\description{
New function that gives tidy data
}

