% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nltt_values.R
\name{get_nltt_values}
\alias{get_nltt_values}
\title{Collects the nLTT values of all phylogenies in the melted/uncast/long form}
\usage{
get_nltt_values(phylogenies, dt)
}
\arguments{
\item{phylogenies}{the phylogenies, supplied as either a list or a multiPhylo object, where the phylogenies are of type 'phylo'}

\item{dt}{The timestep resolution, where 1/dt is the number of points evaluated}
}
\value{
A dataframe of timepoints with the nLTT value of each phylogeny in time
}
\description{
Collects the nLTT values of all phylogenies in the melted/uncast/long form
}

