% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_distributed_normally.R
\name{is_distributed_normally}
\alias{is_distributed_normally}
\title{Sees if the values of a distribution are normally distributed
 using a Shapiro-Wilk normality test}
\usage{
is_distributed_normally(values, p_value = 0.05)
}
\arguments{
\item{values}{the values to check}

\item{p_value}{the p value}
}
\value{
TRUE when it cannot be rejected that the distribution is normal,
   or FALSE when it can be rejected that the distribution is normal
}
\description{
Sees if the values of a distribution are normally distributed
 using a Shapiro-Wilk normality test
}
\examples{
   # Create a normal disribution
   nd <- rnorm(n = 1000, mean = 0.0, sd = 1.0)
   testit::assert(is_distributed_normally(nd))

   # Create a non-normal disribution
   nnd <- runif(n = 1000, min = 0.0, max = 1.0)
   testit::assert(!is_distributed_normally(nnd))
}
\author{
Richel Bilderbeek
}

