% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_on_ci.R, R/is_on_travis.R
\name{is_on_travis}
\alias{is_on_travis}
\title{Determines if the environment is Travis CI}
\usage{
is_on_travis()

is_on_travis()
}
\value{
TRUE if run on Travis CI, FALSE otherwise

TRUE if run on Travis CI, FALSE otherwise
}
\description{
Determines if the environment is Travis CI

Determines if the environment is Travis CI
}
\author{
Richèl J.C. Bilderbeek

Richèl J.C. Bilderbeek
}
