% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_phylogeny.R
\name{is_phylogeny}
\alias{is_phylogeny}
\title{Checks if the input is a phylogeny}
\usage{
is_phylogeny(x)
}
\arguments{
\item{x}{input to be checked}
}
\value{
TRUE or FALSE
}
\description{
Checks if the input is a phylogeny
}
\author{
Richel Bilderbeek
}

