% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_whole_number.R
\name{is_whole_number}
\alias{is_whole_number}
\title{Checks if x is a single, whole-number variable}
\usage{
is_whole_number(x, tolerance = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{the number to check}

\item{tolerance}{the maximum error a number may deviate from a whole number,
before it is labeled as a floating point value}
}
\value{
TRUE or FALSE
}
\description{
Checks if x is a single, whole-number variable
}
\author{
Richel Bilderbeek
}

