% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_rbind_dfs.R
\name{rbind_dfs}
\alias{rbind_dfs}
\title{Combines multiple random-row data frames using rbind.
It is to be compared to rcpp_rbind_dfs, that does the
same using Rcpp}
\usage{
rbind_dfs()
}
\description{
Combines multiple random-row data frames using rbind.
It is to be compared to rcpp_rbind_dfs, that does the
same using Rcpp
}
\author{
Richel Bilderbeek
}

