% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_rbind_dfs.R
\name{rcpp_rbind_dfs}
\alias{rcpp_rbind_dfs}
\title{Combines multiple random-row data frames using rbind.
It is to be compared to rbind_dfs, that does the
same without using Rcpp}
\usage{
rcpp_rbind_dfs()
}
\description{
Combines multiple random-row data frames using rbind.
It is to be compared to rbind_dfs, that does the
same without using Rcpp
}
\author{
Richel Bilderbeek
}

