% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_calc_harm_mean.R
\name{rkt_calc_harm_mean}
\alias{rkt_calc_harm_mean}
\title{Calculate the harmonic mean of a numeric vector}
\usage{
rkt_calc_harm_mean(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the harmonic mean of vector \code{x}
}
\description{
Calculate the harmonic mean of a numeric vector
}
\author{
Richel J.C. Bilderbeek
}
